Disable on-update triggers much more aggressively
Within a single transaction, just disabling the trigger will still cause it to gather up events to fire on, causing issues when it is reenabled in the same transaction. Rather than include a `commit` in a migration (which would break the fact that our migrations are atomic), I decided to just drop and recreate the trigger rather than disable and reenable it. The only difference in effect is that a dropped trigger cannot gather events like a disabled trigger.